added uci basic functions and changed to negamax

+++ src/search/eval.h
@@ -1,4 +1,4 @@
 #include "../misc/defines.h"
 #include "../representation/board.h"
 
-double evaluate(Board &b);
+int evaluate(Board &b);
